Add basic CSS styles
Added a stylesheet used externally by all files. A consistent font
family is used across the site. Each recipe is split into a title,
description, ingredient, and steps sections and each section is targeted
in the stylesheet. The README is also updated to reflect that the
project also incorporates CSS and not just HTML.